+++ extend/waibuMpa/theme/component/widget/form-file.js
@@ -9,7 +9,7 @@ async function formFile () {
     }
 
     build = async () => {
-      await build.call(this, buildFormFile, this.params)
+      await build.call(this, buildFormFile)
     }
   }
 }