+++ extend/waibuMpa/theme/component/widget/form-password.js
@@ -5,7 +5,7 @@ async function formPassword () {
   return class FormPassword extends this.app.baseClass.MpaWidget {
     build = async () => {
       this.params.attr.type = 'password'
-      await build.call(this, buildFormInput, this.params)
+      await build.call(this, buildFormInput)
     }
   }
 }