+++ extend/waibuMpa/theme/component/widget/form-radio-toggle.js
@@ -5,7 +5,7 @@ async function formRadioToggle () {
   return class FormRadioToggle extends this.app.baseClass.MpaWidget {
     build = async () => {
       if (!this.params.attr.label) this.params.attr.label = this.component.req.t('Toggle Radio')
-      await build.call(this, buildFormRadioToggle, this.params)
+      await build.call(this, buildFormRadioToggle)
     }
   }
 }