+++ extend/waibuMpa/theme/component/widget/form-switch.js
@@ -10,7 +10,7 @@ async function formSwitch () {
     }
 
     build = async () => {
-      await build.call(this, buildFormSwitch, this.params)
+      await build.call(this, buildFormSwitch)
     }
   }
 }