+++ extend/waibuMpa/theme/component/widget/form-textarea.js
@@ -4,7 +4,7 @@ import { build } from './form-input.js'
 async function formTextarea () {
   return class FormTextarea extends this.app.baseClass.MpaWidget {
     build = async () => {
-      await build.call(this, buildFormTextarea, this.params)
+      await build.call(this, buildFormTextarea)
     }
   }
 }